\name{colorscales}

\alias{eventFillScale}
\alias{verdictColorScale}
\alias{rateFillScale}
\alias{baseFillScale}
                  
\title{Rariant color scales}
   
\description{

  Color and fill scales used for plotting in Rariant.
  
}
  
\usage{

  eventFillScale()
  verdictColorScale()
  rateFillScale()
  baseFillScale()

}
  
\value{

  A ggbio color or fill scale.

}

\seealso{

  tallyPlot, evidenceHeatmap, plotConfidenceIntervals

}
