\name{Rariant-package}
    
\docType{package}

\alias{Rariant}
\alias{Rariant-package}

\title{Rariant package}

\description{

  The 'Rariant' package offers the framework to identify and
  characterize shifts of variant frequencies in a comparative setting
  from high-throughput short-read sequencing data.  It estimates shifts
  in the non-consensus variant frequency and provides confidence
  estimates that allow for a quantitative assessment of presence or
  absence of variants.  The vignette accompanying the package gives a
  detailed explanation and outlines a typical workflow on real data.
  
}

\seealso{
  
  rariant

  vignette(package = "Rariant")
  
}

\author{

  Julian Gehring, Simon Anders, Bernd Klaus (EMBL Heidelberg)

  Maintainer: Julian Gehring <julian.gehring@embl.de>

}

\examples{

help("rariant")

vignette(package = "Rariant")

}

\keyword{package}
